pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "finance-bot"
version = "0.1.0"
description = "Bot Telegram para registro de despesas pessoais via áudio"
requires-python = ">=3.13"
dependencies = [
    "python-telegram-bot==22.5",
    "pydantic-settings==2.7.0",
    "sqlmodel==0.0.24",
    "alembic==1.13.0",
    "psycopg2-binary==2.9.10",
    "pytest==8.3.4",
    "pytest-asyncio==0.24.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[tool.ruff]
line-length = 100
target-version = "py313"
exclude = [".git", ".venv", "__pycache__", "build", "dist", ".mypy_cache", ".ruff_cache"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*.py" = ["S101"]
"alembic/env.py" = ["E402"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
exclude = ["^build/", "^dist/", "\\.venv/"]

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false
ignore_missing_imports = true
